added annotations to Todo class and changed Id to type Long

+++ src/main/java/com/rest/webservices/RESTfulservicesforAngularToDoList/TodoHardcodedService.java
@@ -10,12 +10,12 @@ import java.util.List;
 public class TodoHardcodedService {
 
     private static List<Todo> todos = new ArrayList();
-    private static int idCounter = 0;
+    private static long idCounter = 0;
 
     static {
-        todos.add(new Todo((long) ++idCounter, "amandaw", "Learn REST", new Date(), false));
-        todos.add(new Todo((long) ++idCounter, "amandaw", "Learn Angular", new Date(), false));
-        todos.add(new Todo((long) ++idCounter, "amandaw", "Get a job", new Date(), false));
+        todos.add(new Todo( ++idCounter, "amandaw", "Learn REST", new Date(), false));
+        todos.add(new Todo( ++idCounter, "amandaw", "Learn Angular", new Date(), false));
+        todos.add(new Todo( ++idCounter, "amandaw", "Get a job", new Date(), false));
     }
 
     public List<Todo> findAll() {
